Update stable series to current Advanced release: 1.25.04
git-svn-id: http://svn.code.sf.net/p/xmlrpc-c/code@2115 98333e67-4a24-44d7-a75c-e53540dd3050

+++ stable/src/double.c
@@ -35,8 +35,8 @@ bufferConcat(buffer * const bufferP,
 
     if (bufferP->bytes) {
         if (bufferP->next >= bufferP->end) {
-            unsigned int const oldSize = bufferP->end - bufferP->bytes;
-            unsigned int const newSize = oldSize + 64;
+            size_t const oldSize = bufferP->end - bufferP->bytes;
+            size_t const newSize = oldSize + 64;
             bufferP->bytes = realloc(bufferP->bytes, newSize);
             bufferP->next = bufferP->bytes + oldSize;
             bufferP->end  = bufferP->bytes + newSize;
